Moves onError handler to store configuration
