+++ src/signals/incident-management/containers/DefaultTextsAdmin/saga.js
@@ -5,30 +5,42 @@ import { authCall, authPostCall } from 'shared/services/api/api';
 
 import { FETCH_DEFAULT_TEXTS, STORE_DEFAULT_TEXTS } from './constants';
 import {
-  fetchDefaultTextsSuccess, fetchDefaultTextsError, storeDefaultTextsSuccess, storeDefaultTextsError,
+  fetchDefaultTextsSuccess,
+  fetchDefaultTextsError,
+  storeDefaultTextsSuccess,
+  storeDefaultTextsError,
 } from './actions';
 
 export function* fetchDefaultTexts(action) {
-  const requestURL = `${CONFIGURATION.API_ROOT}signals/v1/private/terms/categories`;
   try {
     const payload = action.payload;
-    const result = yield call(authCall, `${requestURL}/${payload.main_slug}/sub_categories/${payload.sub_slug}/status-message-templates`);
+    const result = yield call(
+      authCall,
+      `${CONFIGURATION.TERMS_ENDPOINT}${payload.main_slug}/sub_categories/${payload.sub_slug}/status-message-templates`
+    );
+
     const found = result.find(item => item.state === payload.state);
+
     yield put(fetchDefaultTextsSuccess((found && found.templates) || []));
   } catch (error) {
     yield put(fetchDefaultTextsError(error));
+    throw error;
   }
 }
 
 export function* storeDefaultTexts(action) {
-  const requestURL = `${CONFIGURATION.API_ROOT}signals/v1/private/terms/categories`;
   try {
     const payload = action.payload;
-    const result = yield call(authPostCall, `${requestURL}/${payload.main_slug}/sub_categories/${payload.sub_slug}/status-message-templates`, [payload.post]);
+    const result = yield call(
+      authPostCall,
+      `${CONFIGURATION.TERMS_ENDPOINT}${payload.main_slug}/sub_categories/${payload.sub_slug}/status-message-templates`,
+      [payload.post]
+    );
     const found = result.find(item => item.state === payload.post.state);
     yield put(storeDefaultTextsSuccess((found && found.templates) || []));
   } catch (error) {
     yield put(storeDefaultTextsError(error));
+    throw(error);
   }
 }
